+++ lib/src/localization/localization_constant.dart
@@ -1,6 +1,6 @@
 import 'dart:io';
 
-import 'package:pocketlist/src/Shared_Prefs/Prefrecias_user.dart';
+import 'package:pocketlist/src/Shared_Prefs/Preferencias_user.dart';
 import 'package:flutter/material.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:pocketlist/src/localization/localization.dart';
