+++ lib/src/pages/home_page.dart
@@ -1,6 +1,6 @@
 // import 'package:pocketlist/src/pages/New-List/newList.dart';
 // import 'package:flutter/material.dart';
-// import 'package:pocketlist/src/Shared_Prefs/Prefrecias_user.dart';
+// import 'package:pocketlist/src/Shared_Prefs/Preferencias_user.dart';
 // import 'package:pocketlist/src/pages/list_page.dart';
 
 // import 'package:pocketlist/src/utils/utils.dart' as utils;
@@ -51,7 +51,7 @@
 // }
 
 import 'package:flutter/material.dart';
-import 'package:pocketlist/src/Shared_Prefs/Prefrecias_user.dart';
+import 'package:pocketlist/src/Shared_Prefs/Preferencias_user.dart';
 import 'package:pocketlist/src/pages/New-List/newList.dart';
 import 'package:pocketlist/src/pages/list_page.dart';
 import 'package:pocketlist/src/localization/localization_constant.dart';
