+++ lib/src/widgets/Menu_widget.dart
@@ -1,4 +1,4 @@
-import 'package:pocketlist/src/Shared_Prefs/Prefrecias_user.dart';
+import 'package:pocketlist/src/Shared_Prefs/Preferencias_user.dart';
 import 'package:pocketlist/src/pages/New-List/newList.dart';
 import 'package:pocketlist/src/utils/package_info_mixin.dart';
